+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Split `Error` into `Error` and `ParserError`.
+
 ### Removed
 - `failure` dependency.
 - Unused error types.